organize docstrings into module and submodules (#414)
* moves doscstrings to folder

* moves class substitution to own file

* separates docstrings into submodules

* fixes codestyle
