+++ feature_engine/_docstrings/class_inputs.py
@@ -1,0 +1,15 @@
+_variables_numerical_docstring = """variables: list, default=None
+        The list of numerical variables to transform. If None, the transformer will
+        automatically find and select all numerical variables.
+    """.rstrip()
+
+_drop_original_docstring = """drop_original: bool, default=False
+        If True, the original variables to transform will be dropped from the dataframe.
+    """.rstrip()
+
+_missing_values_docstring = """missing_values: string, default='raise'
+        Indicates if missing values should be ignored or raised. If 'raise' the
+        transformer will return an error if the the datasets to `fit` or `transform`
+        contain missing values. If 'ignore', missing data will be ignored when learning
+        parameters or performing the transformation.
+        """.rstrip()
