+++ feature_engine/_docstrings/fit_attributes.py
@@ -1,0 +1,11 @@
+_variables_attribute_docstring = """variables_:
+        The group of variables that will be transformed.
+        """.rstrip()
+
+_feature_names_in_docstring = """feature_names_in_:
+        List with the names of features seen during `fit`.
+        """.rstrip()
+
+_n_features_in_docstring = """n_features_in_:
+        The number of features in the train set used in fit.
+        """.rstrip()
