+++ feature_engine/_docstrings/methods.py
@@ -1,0 +1,20 @@
+_fit_not_learn_docstring = """fit:
+        This transformer does not learn parameters.
+        """.rstrip()
+
+_fit_transform_docstring = """fit_transform:
+        Fit to data, then transform it.
+
+    get_feature_names_out:
+        Get output feature names for transformation.
+
+    get_params:
+        Get parameters for this estimator.
+
+    set_params:
+        Set the parameters of this estimator.
+        """.rstrip()
+
+_inverse_transform_docstring = """inverse_transform:
+        Convert the data back to the original representation.
+        """.rstrip()
