+++ feature_engine/creation/combine_with_reference_feature.py
@@ -11,15 +11,19 @@ from feature_engine.dataframe_checks import (
     _check_X_matches_training_df,
     check_X,
 )
-from feature_engine.docstrings import (
-    Substitution,
-    _drop_original_docstring,
-    _feature_names_in_docstring,
+from feature_engine._docstrings.methods import (
     _fit_not_learn_docstring,
     _fit_transform_docstring,
-    _missing_values_docstring,
+)
+from feature_engine._docstrings.fit_attributes import (
+    _feature_names_in_docstring,
     _n_features_in_docstring,
 )
+from feature_engine._docstrings.class_inputs import (
+    _drop_original_docstring,
+    _missing_values_docstring,
+)
+from feature_engine._docstrings.substitute import Substitution
 from feature_engine.tags import _return_tags
 from feature_engine.variable_manipulation import _find_or_check_numerical_variables
