+++ feature_engine/creation/relative_features.py
@@ -4,16 +4,20 @@ import pandas as pd
 from sklearn.utils.validation import check_is_fitted
 
 from feature_engine.creation.base_creation import BaseCreation
-from feature_engine.docstrings import (
-    Substitution,
-    _drop_original_docstring,
-    _feature_names_in_docstring,
+from feature_engine._docstrings.methods import (
     _fit_not_learn_docstring,
     _fit_transform_docstring,
-    _missing_values_docstring,
-    _n_features_in_docstring,
-    _variables_numerical_docstring,
 )
+from feature_engine._docstrings.fit_attributes import (
+    _feature_names_in_docstring,
+    _n_features_in_docstring,
+)
+from feature_engine._docstrings.class_inputs import (
+    _variables_numerical_docstring,
+    _drop_original_docstring,
+    _missing_values_docstring,
+)
+from feature_engine._docstrings.substitute import Substitution
 from feature_engine.variable_manipulation import _find_or_check_numerical_variables
 
 _PERMITTED_FUNCTIONS = [
