+++ feature_engine/discretisation/arbitrary.py
@@ -7,14 +7,16 @@ from typing import Dict, List, Optional, Union
 import pandas as pd
 
 from feature_engine.discretisation.base_discretiser import BaseDiscretiser
-from feature_engine.docstrings import (
-    Substitution,
-    _feature_names_in_docstring,
+from feature_engine._docstrings.methods import (
     _fit_not_learn_docstring,
     _fit_transform_docstring,
-    _n_features_in_docstring,
-    _variables_attribute_docstring,
 )
+from feature_engine._docstrings.fit_attributes import (
+    _variables_attribute_docstring,
+    _feature_names_in_docstring,
+    _n_features_in_docstring,
+)
+from feature_engine._docstrings.substitute import Substitution
 from feature_engine.tags import _return_tags
