+++ feature_engine/encoding/one_hot.py
@@ -7,13 +7,13 @@ import numpy as np
 import pandas as pd
 from sklearn.utils.validation import check_is_fitted
 
-from feature_engine.docstrings import (
-    Substitution,
-    _feature_names_in_docstring,
-    _fit_transform_docstring,
-    _n_features_in_docstring,
+from feature_engine._docstrings.methods import _fit_transform_docstring
+from feature_engine._docstrings.fit_attributes import (
     _variables_attribute_docstring,
+    _feature_names_in_docstring,
+    _n_features_in_docstring,
 )
+from feature_engine._docstrings.substitute import Substitution
 from feature_engine.encoding._docstrings import (
     _ignore_format_docstring,
     _variables_docstring,
