+++ feature_engine/outliers/artbitrary.py
@@ -11,15 +11,17 @@ from feature_engine.dataframe_checks import (
     _check_contains_na,
     check_X,
 )
-from feature_engine.docstrings import (
-    Substitution,
-    _feature_names_in_docstring,
+from feature_engine._docstrings.methods import (
     _fit_not_learn_docstring,
     _fit_transform_docstring,
-    _missing_values_docstring,
-    _n_features_in_docstring,
-    _variables_attribute_docstring,
 )
+from feature_engine._docstrings.fit_attributes import (
+    _variables_attribute_docstring,
+    _feature_names_in_docstring,
+    _n_features_in_docstring,
+)
+from feature_engine._docstrings.class_inputs import _missing_values_docstring
+from feature_engine._docstrings.substitute import Substitution
 from feature_engine.outliers.base_outlier import BaseOutlier
 from feature_engine.parameter_checks import _define_numerical_dict
 from feature_engine.tags import _return_tags
