+++ feature_engine/selection/drop_psi_features.py
@@ -14,12 +14,12 @@ from feature_engine.discretisation import (
     EqualFrequencyDiscretiser,
     EqualWidthDiscretiser,
 )
-from feature_engine.docstrings import (
-    Substitution,
+from feature_engine._docstrings.methods import _fit_transform_docstring
+from feature_engine._docstrings.fit_attributes import (
     _feature_names_in_docstring,
-    _fit_transform_docstring,
     _n_features_in_docstring,
 )
+from feature_engine._docstrings.substitute import Substitution
 from feature_engine.selection._docstring import (
     _variables_attribute_docstring,
     _variables_numerical_docstring,
