+++ feature_engine/selection/single_feature_performance.py
@@ -5,12 +5,12 @@ import pandas as pd
 from sklearn.model_selection import cross_validate
 
 from feature_engine.dataframe_checks import check_X
-from feature_engine.docstrings import (
-    Substitution,
+from feature_engine._docstrings.methods import _fit_transform_docstring
+from feature_engine._docstrings.fit_attributes import (
     _feature_names_in_docstring,
-    _fit_transform_docstring,
     _n_features_in_docstring,
 )
+from feature_engine._docstrings.substitute import Substitution
 from feature_engine.selection._docstring import (
     _cv_docstring,
     _estimator_docstring,
