+++ feature_engine/selection/smart_correlation_selection.py
@@ -8,12 +8,12 @@ from feature_engine.dataframe_checks import (
     _check_contains_na,
     check_X,
 )
-from feature_engine.docstrings import (
-    Substitution,
+from feature_engine._docstrings.methods import _fit_transform_docstring
+from feature_engine._docstrings.fit_attributes import (
     _feature_names_in_docstring,
-    _fit_transform_docstring,
     _n_features_in_docstring,
 )
+from feature_engine._docstrings.substitute import Substitution
 from feature_engine.selection._docstring import (
     _cv_docstring,
     _estimator_docstring,
