+++ feature_engine/timeseries/forecasting/__init__.py
@@ -4,8 +4,4 @@ from .expanding_window_features import ExpandingWindowFeatures
 from .lag_features import LagFeatures
 from .window_features import WindowFeatures
 
-__all__ = [
-    "LagFeatures",
-    "WindowFeatures",
-    "ExpandingWindowFeatures"
-]
+__all__ = ["LagFeatures", "WindowFeatures", "ExpandingWindowFeatures"]
