+++ feature_engine/timeseries/forecasting/base_forecast_transformers.py
@@ -10,14 +10,16 @@ from feature_engine.dataframe_checks import (
     _check_X_matches_training_df,
     check_X,
 )
-from feature_engine.docstrings import (
-    Substitution,
-    _drop_original_docstring,
+from feature_engine._docstrings.methods import _fit_not_learn_docstring
+from feature_engine._docstrings.fit_attributes import (
     _feature_names_in_docstring,
-    _fit_not_learn_docstring,
-    _missing_values_docstring,
     _n_features_in_docstring,
 )
+from feature_engine._docstrings.class_inputs import (
+    _drop_original_docstring,
+    _missing_values_docstring,
+)
+from feature_engine._docstrings.substitute import Substitution
 from feature_engine.tags import _return_tags
 from feature_engine.variable_manipulation import (
     _check_input_parameter_variables,
